+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Add internal GraphQL interface to Shopify's GraphQL API
 
 ## 0.5.1